requirements.txt
slackclient==2.9.3
python-decouple==3.4
slack-bolt==1.6.0
